Add files via upload
This is an othello AI project. My teacher built the main game to save on time, but the kje2115 was created by me. To run the base game, run only othello_gui.py in python 3. To run the AI, run the file with one of the AI files as an argument (ex: python3 othello_gui.py kje2115_ai.py). The AI is a simple minimax algorithm, with some extra heuristics including going for corners and taking up as much space as possible.